Create three arrays and start data getting
There are now three arrays available: ObjArr (meant for arrays of arrays), intArr and BoolArr. Also, the data is fetched in lines thanks to the while cycle, and what now follows is to figure out a way to get the data in an array of arrays (may need to implement __iter__ for the arrays once an array of stations is passed instead of a single one.