feat(/): data array and data structures
Manage to retrieve all the relevant info from the CSV file, that is, all the rows that correspond to info about stations in any of the zones through which a bus route goes. In the future, this will be optimized to make sure it also filters the relevant stations within the zones. On the other hand, the name of the data structures' files are changed, to comply with the standard of naming a module containing an object with the same name as the object but in camelcase. In addition, I make some changes to the object array and create an array of strings.